* Refactored ezPropertyFlags: StandardType|IsEnum|Bitflags|Class now match exactly the type flags equivalents. One and only one must be set.   Const|Reference|Pointer are additions, any combinations is possible except for *& which I couldn't be bothered to implement.   Removed Constant as it was redundant.   * Lots of refactoring and cleanup goes along with this to makes things more readable, hope I didn't break anything. * Member function reflection: Arguments are provided as ezVariant array.   /// The following rules apply for both arguments and return value:   /// Any standard type (bool, ezVec3, ezVariant...) must be provided by value, even if it is a pointer to one.   /// enum and bitflags are supported if ezEnum / ezBitflags is used.   /// Out values (&, *) are written back to the variant they were read from.   /// Any complex class is provided by pointer, regardless of whether it is a pointer or not.   /// An invalid variant is equal to a nullptr, except for if the parameter is of type ezVariant, in which case it is impossible to pass along a nullptr.   See unit test for examples. * TODO: static and constructor functions, macros for easy definition. * natvis update for ezVariant. * Can now compare and set ezVariant with const char*

+++ Code/Engine/Foundation/Serialization/Implementation/RttiConverterReader.cpp
@@ -56,11 +56,7 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
     {
       ezAbstractMemberProperty* pSpecific = static_cast<ezAbstractMemberProperty*>(pProp);
 
-      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType | ezPropertyFlags::IsEnum | ezPropertyFlags::Bitflags))
-      {
-        ezReflectionUtils::SetMemberPropertyValue(pSpecific, pObject, pSource->m_Value);
-      }
-      else if (pProp->GetFlags().IsSet(ezPropertyFlags::Pointer))
+      if (pProp->GetFlags().IsSet(ezPropertyFlags::Pointer))
       {
         if (!pSource->m_Value.IsA<ezUuid>())
           return;
@@ -95,28 +91,35 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
       else
       {
-        if (!pSource->m_Value.IsA<ezUuid>())
-          return;
-
-        void* pDirectPtr = pSpecific->GetPropertyPointer(pObject);
-        bool bDelete = false;
-        const ezUuid sourceGuid = pSource->m_Value.Get<ezUuid>();
-
-        if (pDirectPtr == nullptr)
+        if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType | ezPropertyFlags::IsEnum | ezPropertyFlags::Bitflags))
         {
-          bDelete = true;
-          pDirectPtr = m_pContext->CreateObject(sourceGuid, pPropType);
+          ezReflectionUtils::SetMemberPropertyValue(pSpecific, pObject, pSource->m_Value);
         }
-
-        auto* pNode = m_pGraph->GetNode(sourceGuid);
-        EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-
-        ApplyPropertiesToObject(pNode, pPropType, pDirectPtr);
-
-        if (bDelete)
+        else if (pProp->GetFlags().IsSet(ezPropertyFlags::Class))
         {
-          pSpecific->SetValuePtr(pObject, pDirectPtr);
-          m_pContext->DeleteObject(sourceGuid);
+          if (!pSource->m_Value.IsA<ezUuid>())
+            return;
+
+          void* pDirectPtr = pSpecific->GetPropertyPointer(pObject);
+          bool bDelete = false;
+          const ezUuid sourceGuid = pSource->m_Value.Get<ezUuid>();
+
+          if (pDirectPtr == nullptr)
+          {
+            bDelete = true;
+            pDirectPtr = m_pContext->CreateObject(sourceGuid, pPropType);
+          }
+
+          auto* pNode = m_pGraph->GetNode(sourceGuid);
+          EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+
+          ApplyPropertiesToObject(pNode, pPropType, pDirectPtr);
+
+          if (bDelete)
+          {
+            pSpecific->SetValuePtr(pObject, pDirectPtr);
+            m_pContext->DeleteObject(sourceGuid);
+          }
         }
       }
     }
@@ -142,14 +145,7 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
 
       pSpecific->SetCount(pObject, array.GetCount());
-      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
-      {
-        for (ezUInt32 i = 0; i < array.GetCount(); ++i)
-        {
-          ezReflectionUtils::SetArrayPropertyValue(pSpecific, pObject, i, array[i]);
-        }
-      }
-      else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
+      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
       {
         for (ezUInt32 i = 0; i < array.GetCount(); ++i)
         {
@@ -159,13 +155,16 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
           void* pRefrencedObject = nullptr;
           if (pProp->GetFlags().IsSet(ezPropertyFlags::PointerOwner))
           {
-            auto* pNode = m_pGraph->GetNode(guid);
-            EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-            pRefrencedObject = CreateObjectFromNode(pNode);
-            if (pRefrencedObject == nullptr)
+            if (guid.IsValid())
             {
-              ezLog::Error("Failed to set array property '{0}' element, type could not be created!", pProp->GetPropertyName());
-              continue;
+              auto* pNode = m_pGraph->GetNode(guid);
+              EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+              pRefrencedObject = CreateObjectFromNode(pNode);
+              if (pRefrencedObject == nullptr)
+              {
+                ezLog::Error("Failed to set array property '{0}' element, type could not be created!", pProp->GetPropertyName());
+                continue;
+              }
             }
           }
           else
@@ -177,24 +176,34 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
       else
       {
-        ezUuid temp;
-        temp.CreateNewUuid();
-        void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
-
-        for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+        if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
         {
-          if (!array[i].IsA<ezUuid>())
-            continue;
-
-          const ezUuid sourceGuid = array[i].Get<ezUuid>();
-          auto* pNode = m_pGraph->GetNode(sourceGuid);
-          EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-
-          ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
-          pSpecific->SetValue(pObject, i, pValuePtr);
+          for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+          {
+            ezReflectionUtils::SetArrayPropertyValue(pSpecific, pObject, i, array[i]);
+          }
         }
+        else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Class))
+        {
+          ezUuid temp;
+          temp.CreateNewUuid();
+          void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
 
-        m_pContext->DeleteObject(temp);
+          for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+          {
+            if (!array[i].IsA<ezUuid>())
+              continue;
+
+            const ezUuid sourceGuid = array[i].Get<ezUuid>();
+            auto* pNode = m_pGraph->GetNode(sourceGuid);
+            EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+
+            ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
+            pSpecific->SetValue(pObject, i, pValuePtr);
+          }
+
+          m_pContext->DeleteObject(temp);
+        }
       }
     }
     break;
@@ -221,14 +230,8 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
 
       pSpecific->Clear(pObject);
-      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
-      {
-        for (ezUInt32 i = 0; i < array.GetCount(); ++i)
-        {
-          ezReflectionUtils::InsertSetPropertyValue(pSpecific, pObject, array[i]);
-        }
-      }
-      else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
+
+      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
       {
         for (ezUInt32 i = 0; i < array.GetCount(); ++i)
         {
@@ -257,24 +260,34 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
       else
       {
-        ezUuid temp;
-        temp.CreateNewUuid();
-        void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
-
-        for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+        if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
         {
-          if (!array[i].IsA<ezUuid>())
-            continue;
-
-          const ezUuid sourceGuid = array[i].Get<ezUuid>();
-          auto* pNode = m_pGraph->GetNode(sourceGuid);
-          EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-
-          ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
-          pSpecific->Insert(pObject, pValuePtr);
+          for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+          {
+            ezReflectionUtils::InsertSetPropertyValue(pSpecific, pObject, array[i]);
+          }
         }
+        else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Class))
+        {
+          ezUuid temp;
+          temp.CreateNewUuid();
+          void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
 
-        m_pContext->DeleteObject(temp);
+          for (ezUInt32 i = 0; i < array.GetCount(); ++i)
+          {
+            if (!array[i].IsA<ezUuid>())
+              continue;
+
+            const ezUuid sourceGuid = array[i].Get<ezUuid>();
+            auto* pNode = m_pGraph->GetNode(sourceGuid);
+            EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+
+            ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
+            pSpecific->Insert(pObject, pValuePtr);
+          }
+
+          m_pContext->DeleteObject(temp);
+        }
       }
     }
     break;
@@ -302,14 +315,8 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
 
       pSpecific->Clear(pObject);
-      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
-      {
-        for (auto it = dict.GetIterator(); it.IsValid(); ++it)
-        {
-          ezReflectionUtils::SetMapPropertyValue(pSpecific, pObject, it.Key(), it.Value());
-        }
-      }
-      else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
+
+      if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Pointer))
       {
         for (auto it = dict.GetIterator(); it.IsValid(); ++it)
         {
@@ -320,13 +327,16 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
           void* pRefrencedObject = nullptr;
           if (pProp->GetFlags().IsSet(ezPropertyFlags::PointerOwner))
           {
-            auto* pNode = m_pGraph->GetNode(guid);
-            EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-            pRefrencedObject = CreateObjectFromNode(pNode);
-            if (pRefrencedObject == nullptr)
+            if (guid.IsValid())
             {
-              ezLog::Error("Failed to insert set element in to property '{0}', type could not be created!", pProp->GetPropertyName());
-              continue;
+              auto* pNode = m_pGraph->GetNode(guid);
+              EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+              pRefrencedObject = CreateObjectFromNode(pNode);
+              if (pRefrencedObject == nullptr)
+              {
+                ezLog::Error("Failed to insert set element in to property '{0}', type could not be created!", pProp->GetPropertyName());
+                continue;
+              }
             }
           }
           else
@@ -338,24 +348,34 @@ void ezRttiConverterReader::ApplyProperty(void* pObject, ezAbstractProperty* pPr
       }
       else
       {
-        ezUuid temp;
-        temp.CreateNewUuid();
-        void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
-
-        for (auto it = dict.GetIterator(); it.IsValid(); ++it)
+        if (pProp->GetFlags().IsAnySet(ezPropertyFlags::StandardType))
         {
-          if (!it.Value().IsA<ezUuid>())
-            continue;
-
-          const ezUuid sourceGuid = it.Value().Get<ezUuid>();
-          auto* pNode = m_pGraph->GetNode(sourceGuid);
-          EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
-
-          ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
-          pSpecific->Insert(pObject, it.Key(), pValuePtr);
+          for (auto it = dict.GetIterator(); it.IsValid(); ++it)
+          {
+            ezReflectionUtils::SetMapPropertyValue(pSpecific, pObject, it.Key(), it.Value());
+          }
         }
+        else if (pProp->GetFlags().IsAnySet(ezPropertyFlags::Class))
+        {
+          ezUuid temp;
+          temp.CreateNewUuid();
+          void* pValuePtr = m_pContext->CreateObject(temp, pPropType);
 
-        m_pContext->DeleteObject(temp);
+          for (auto it = dict.GetIterator(); it.IsValid(); ++it)
+          {
+            if (!it.Value().IsA<ezUuid>())
+              continue;
+
+            const ezUuid sourceGuid = it.Value().Get<ezUuid>();
+            auto* pNode = m_pGraph->GetNode(sourceGuid);
+            EZ_ASSERT_DEV(pNode != nullptr, "node must exist");
+
+            ApplyPropertiesToObject(pNode, pPropType, pValuePtr);
+            pSpecific->Insert(pObject, it.Key(), pValuePtr);
+          }
+
+          m_pContext->DeleteObject(temp);
+        }
       }
     }
     break;
